requirements.txt
python-dotenv
gradio
langchain
langchain_core
langchain_community
langchain_huggingface
langchain_ollama
langgraph
langfuse